+++ src/extension-snippets.js
@@ -1,15 +1,21 @@
 /**
- * Extension-specific snippets.
+ * Snippets for a particular extension. An Extension can have snippets declared in
+ * different files. One snippet file can be associated with 1 or more languages.
  */
 class ExtensionSnippets {
   constructor(name, displayName, publisher) {
     this.name = name;
     this.displayName = displayName;
-		this.publisher = publisher;
-		this.id = `${publisher}.${name}`;
+    this.publisher = publisher;
+    this.id = `${publisher}.${name}`;
     this.snippets = [];
   }
 
+  /**
+   *
+   * @param {String} language The programming language associated with the file.
+   * @param {String} path The absolute file path for the snippet file.
+   */
   addSnippetsFile(language, path) {
     let matched = false;
     this.snippets.forEach((snippet, index) => {