+++ webpack.config.js
@@ -1,10 +1,22 @@
+/* eslint-disable node/no-unpublished-require */
 const path = require("path");
+const MiniCssExtractPlugin = require("mini-css-extract-plugin");
+const OptimizeCSSAssetsPlugin = require("optimize-css-assets-webpack-plugin");
+const TerserPlugin = require("terser-webpack-plugin");
 
 const config = {
-  target: "node", // vscode extensions run in a Node.js-context 📖 -> https://webpack.js.org/configuration/node/
-  entry: "./src/extension.js",
+  target: "node",
+  entry: { main: "./src/extension.js" },
+  optimization: {
+    minimizer: [new OptimizeCSSAssetsPlugin(), new TerserPlugin()],
+  },
+  plugins: [new MiniCssExtractPlugin({ filename: "styles.minified.css" })],
+  module: {
+    rules: [
+      { test: /\.css$/, use: [MiniCssExtractPlugin.loader, "css-loader"] },
+    ],
+  },
   output: {
-    // the bundle is stored in the 'dist' folder (check package.json), 📖 -> https://webpack.js.org/configuration/output/
     path: path.resolve(__dirname, "dist"),
     filename: "extension.js",
     libraryTarget: "commonjs2",
@@ -12,7 +24,7 @@ const config = {
   },
   devtool: "source-map",
   externals: {
-    vscode: "commonjs vscode", // the vscode-module is created on-the-fly and must be excluded. 📖 -> https://webpack.js.org/configuration/externals/
+    vscode: "commonjs vscode", // the vscode-module is created on-the-fly and must be excluded.
   },
 };
 module.exports = config;