+++ CHANGELOG.md
@@ -5,7 +5,22 @@ All notable changes to this project are documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.2.0] - 2020-09-02
+## [0.3.0] - 2020-09-04
+
+### Added
+
+- Go to top link.
+
+### Changed
+
+- Changed logo.
+- Changed z-index on h4 so that it slides under h3 and h2 on scroll.
+- Added node engines to package.json to stop ES Lint reporting unsupported features errors.
+- Update rules ES Lint config.
+- Adding formatting of snippets to ensure data is consistent and that the body can be shown on a webpage without formatting issues. Added *snippet.js* which will ensure that *description* is set to blank if it is undefined, and that *body* is always an array. Added *escapeBody()* to *Formatter.js* that will escape markup in the *body* array and concatenate it into a well-formatted string for use in the webview.
+- Changed webpack config to minify the stylesheet.
+
+## [0.2.0] - 2020-09-01
 
 ### Added